* commit '1923992f1e32f8e662d0e9f198be205767072ee1':
  AGORA-5816: remove branch from version
  AGORA-5816: add Timer metric for consumption durations
  AGORA-5816: monitor branched snapshot bersion
  AGORA-5816: move mercury config classes to a separate packages
  AGORA-5816: add mercury metrics handler
  AGORA-5816: move mercury publisher config classes to a separate package

+++ mercury/src/main/java/com/odesk/agora/mercury/consumer/TopicQueueListener.java
@@ -5,12 +5,12 @@ import com.amazonaws.services.sqs.buffered.AmazonSQSBufferedAsyncClient;
 import com.amazonaws.services.sqs.model.DeleteMessageRequest;
 import com.amazonaws.services.sqs.model.Message;
 import com.amazonaws.services.sqs.model.ReceiveMessageRequest;
-import com.amazonaws.services.sqs.model.ReceiveMessageResult;
 import com.amazonaws.util.json.Jackson;
 import com.odesk.agora.mercury.MercuryMessage;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.util.List;
 import java.util.concurrent.Executor;
 import java.util.function.Consumer;
 
@@ -24,6 +24,8 @@ import java.util.function.Consumer;
  * Not intended to be used by end-users.
  */
 public class TopicQueueListener implements Runnable {
+    public final static int MAX_NUMBER_OF_MESSAGES_PER_POLL = 10; //no needs to override this setting
+
     private final Logger logger;
 
     private final String topicName;
@@ -32,24 +34,26 @@ public class TopicQueueListener implements Runnable {
 
     private final AmazonSQSBufferedAsyncClient sqsClient;
     private final Executor consumptionExecutor;
+    private final ConsumerMetricsHandler metricsHandler;
 
     private final String topicNameForLogging;
     private final ReceiveMessageRequest receiveMessageRequest;
     private final AsyncHandler<DeleteMessageRequest, Void> deletionAsyncHandler;
 
     public TopicQueueListener(String topicName, String queueUrl, boolean isDLQ,
-                              AmazonSQSBufferedAsyncClient sqsClient, Executor consumptionExecutor) {
+                              AmazonSQSBufferedAsyncClient sqsClient, Executor consumptionExecutor, ConsumerMetricsHandler metricsHandler) {
         this.sqsClient = sqsClient;
         this.queueUrl = queueUrl;
         this.isDLQ = isDLQ;
         this.topicName = topicName;
         this.consumptionExecutor = consumptionExecutor;
+        this.metricsHandler = metricsHandler;
 
         this.topicNameForLogging = topicName + (isDLQ ? "-DLQ" : "");
 
         logger = LoggerFactory.getLogger(TopicQueueListener.class.getName() + "-" + topicNameForLogging);
 
-        receiveMessageRequest = new ReceiveMessageRequest().withQueueUrl(queueUrl).withMaxNumberOfMessages(10);
+        receiveMessageRequest = new ReceiveMessageRequest().withQueueUrl(queueUrl).withMaxNumberOfMessages(MAX_NUMBER_OF_MESSAGES_PER_POLL);
 
         deletionAsyncHandler = new AsyncHandler<DeleteMessageRequest, Void>() {
             @Override
@@ -67,12 +71,33 @@ public class TopicQueueListener implements Runnable {
         if(MercuryConsumers.getConsumerForTopic(topicName, isDLQ) == null) {
             logger.warn("No consumer registered for {} yet. Skip SQS fetching.", topicNameForLogging);
         } else {
-            ReceiveMessageResult pollResult = sqsClient.receiveMessage(receiveMessageRequest);
+            List<Message> pollingResult;
 
-            if(! pollResult.getMessages().isEmpty()) {
-                logger.info("Received {} messages", pollResult.getMessages().size());
+            try {
+                pollingResult = sqsClient.receiveMessage(receiveMessageRequest).getMessages();
+                if(metricsHandler != null) {
+                    metricsHandler.handlePollingSuccess(topicNameForLogging, pollingResult.size());
+                }
+            } catch(Throwable t) {
+                logger.warn("Polling error", t);
+                if(metricsHandler != null) {
+                    metricsHandler.handlePollingFail(topicNameForLogging);
+                }
+                throw t;
+            }
 
-                for(Message message : pollResult.getMessages()) {
+            if(pollingResult.isEmpty()) {
+                if(metricsHandler != null) {
+                    metricsHandler.handleEmptyPollingResult(topicNameForLogging);
+                }
+            } else {
+                logger.info("Received {} messages", pollingResult.size());
+
+                if((pollingResult.size() == MAX_NUMBER_OF_MESSAGES_PER_POLL) && (metricsHandler != null)) {
+                    metricsHandler.handleFullPollingResult(topicNameForLogging);
+                }
+
+                for(Message message : pollingResult) {
                     consumptionExecutor.execute(new ConsumptionJob(message));
                 };
             }
@@ -82,9 +107,11 @@ public class TopicQueueListener implements Runnable {
 
     public class ConsumptionJob implements Runnable {
         private final Message message;
+        private final long deliveryTime;
 
         public ConsumptionJob(Message message) {
             this.message = message;
+            this.deliveryTime = System.currentTimeMillis();
         }
 
         public String getTopicName() {
@@ -93,35 +120,68 @@ public class TopicQueueListener implements Runnable {
 
         @Override
         public void run() {
+            boolean processed = false;
+
             Consumer<Message> plainSQSConsumer = PlainSQSConsumers.getConsumerForTopic(topicName, isDLQ);
 
             if(plainSQSConsumer != null) {
-                try {
-                    logger.debug("Processing plain SQS message {}", message);
-                    plainSQSConsumer.accept(message);
-                } catch (Throwable t) {
-                    logger.error("Can't process plain SQS message", t);
-                    return;
-                }
+                processed = processAsPlainSQSMessage(plainSQSConsumer);
             } else {
-                Consumer<MercuryMessage> consumer = MercuryConsumers.getConsumerForTopic(topicName, isDLQ);
+                Consumer<MercuryMessage> mercuryConsumer = MercuryConsumers.getConsumerForTopic(topicName, isDLQ);
 
-                if(consumer == null) {
+                if(mercuryConsumer == null) {
                     logger.error("No consumer found for {}. Skip message processing.", topicNameForLogging);
-                    return;
-                }
-
-                try {
-                    MercuryMessage mercuryMessage = Jackson.fromJsonString(message.getBody(), MercuryMessage.class);
-                    logger.debug("Processing message {}", mercuryMessage);
-                    consumer.accept(mercuryMessage);
-                } catch (Throwable t) {
-                    logger.error("Can't process Mercury message", t);
-                    return;
+                } else {
+                    processed = processAsMercuryMessage(mercuryConsumer);
                 }
             }
 
-            sqsClient.deleteMessageAsync(new DeleteMessageRequest(queueUrl, message.getReceiptHandle()), deletionAsyncHandler);
+            if(processed) {
+                sqsClient.deleteMessageAsync(new DeleteMessageRequest(queueUrl, message.getReceiptHandle()), deletionAsyncHandler);
+            }
+        }
+
+        private boolean processAsPlainSQSMessage(Consumer<Message> plainSQSConsumer) {
+            try {
+                logger.debug("Processing plain SQS message {}", message);
+                plainSQSConsumer.accept(message);
+                handleMetricConsumptionSuccess();
+                return true;
+            } catch (Throwable t) {
+                logger.error("Can't process plain SQS message", t);
+                handleMetricConsumptionFail();
+                return false;
+            }
+        }
+
+        private boolean processAsMercuryMessage(Consumer<MercuryMessage> mercuryConsumer) {
+            try {
+                MercuryMessage mercuryMessage = Jackson.fromJsonString(message.getBody(), MercuryMessage.class);
+                if(metricsHandler != null && !isDLQ) {
+                    metricsHandler.handleDeliveryLatency(topicNameForLogging, deliveryTime - mercuryMessage.getTimestamp().getTime());
+                }
+
+                logger.debug("Processing message {}", mercuryMessage);
+                mercuryConsumer.accept(mercuryMessage);
+                handleMetricConsumptionSuccess();
+                return true;
+            } catch (Throwable t) {
+                logger.error("Can't process Mercury message", t);
+                handleMetricConsumptionFail();
+                return false;
+            }
+        }
+
+        private void handleMetricConsumptionFail() {
+            if(metricsHandler != null) {
+                metricsHandler.handleConsumptionFail(topicNameForLogging);
+            }
+        }
+
+        private void handleMetricConsumptionSuccess() {
+            if(metricsHandler != null) {
+                metricsHandler.handleConsumptionDuration(topicNameForLogging, System.currentTimeMillis() - deliveryTime);
+            }
         }
     }
 }
